Remove old tree and fix tests
