--- beacon-chain/cache/depositcache/log.go
@@ -1,5 +1,0 @@
-package depositcache
-
-import "github.com/sirupsen/logrus"
-
-var log = logrus.WithField("prefix", "depositcache")
